reset-ease: correctly mark changed cards for graceful sync
It's not recommended to write directly to the database, because Anki has
its own tracking of card changes (which allows it to do graceful
synchronisation). So simply switch to the recommended flow for updating
cards, and this solves the issue where ease factor changes weren't being
synchronised at all between devices. In addiiton, don't touch cards that
haven't yet been reviewed (those with ease factor 0).

Fixes: AnKingMed/Reset-Ease#3
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>